解决警告

+++ IMiPhone/UserDataProxy.h
@@ -25,7 +25,7 @@
 @property (nonatomic, retain) NSString *mobile;
 @property (nonatomic, retain) NSString *password;
 
-@property (nonatomic) NSInteger showUserInfoUid;
+@property (nonatomic) long long showUserInfoUid;
 @property (nonatomic) NSInteger showUserInfoRleation;
 
 + (UserDataProxy *)sharedProxy;
